moved api calls and fee parsing  to separate file
