+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@googlemaps/js-api-loader": "^1.12.2",
     "axios": "^0.21.1",
+    "dotenv": "^10.0.0",
     "ejs": "^3.1.6",
     "express": "^4.17.1",
     "method-override": "^3.0.0",
